requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
google-genai
google-generativeai>=0.3.0
gradio>=4.0.0
python-dotenv>=1.0.0
duckduckgo-search>=4.1.1
pydantic>=2.0.0
typing-extensions>=4.5.0 
langgraph==0.4.8
IPython==8.37.0